pyproject.toml
lint = ["ruff>=0.0.220"] # MIT License (MIT)

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leakpro"
version = "0.1"
description = "A package for privacy risk analysis"
authors = [{ name = "LeakPro team", email = "johan.ostman@ai.se" }]
readme = "README.md"
license = {file="LICENSE"}
keywords = [
    "Privacy",
    "Risk analysis",
    "Centralized learning",
    "Federated learning",
    "Synthethic data",
    "Machine learning",
]
classifiers = [
    "Natural Language :: English",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

requires-python = '>=3.8,<3.13'

dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "scikit-learn",
  "matplotlib",
  "seaborn",
  "pillow",
  "torch",
  "torchvision",
  "dotmap",
  "loguru",
  "jinja2",
  "tqdm",
  "pyyaml",
  "numba",
  "pydantic",
  "joblib",
  "pytest",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["leakpro*"]

[tool.ruff]
line-length = 130
target-version = "py39"

lint.select = [
  "ANN", # flake8-annotations
  "ARG", # flake8-unused-arguments
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "C90", # mccabe
  "D",   # pydocstyle
  "DTZ", # flake8-datetimez
  "E",   # pycodestyle
  "ERA", # eradicate
  "F",   # Pyflakes
  "I",   # isort
  "N",   # pep8-naming
  "PD",  # pandas-vet
  "PGH", # pygrep-hooks
  "PLC", # Pylint
  "PLE", # Pylint
  "PLR", # Pylint
  "PLW", # Pylint
  "PT",  # flake8-pytest-style
  "Q",   # flake8-quotes
  "RET", # flake8-return
  "S",   # flake8-bandit
  "SIM", # flake8-simplify
  "T20", # flake8-print
  "TID", # flake8-tidy-imports
  "W",   # pycodestyle
]

exclude = [
  ".venv",
  "./tests",
]

lint.ignore = [
  "D401", # non-imperative-mood
  "PD901", # Avoid using df variable name as pd.Dataframe
  "PD011", # pandas-use-of-dot-values
  "PLR0913", # too-many-arguments in function
  "S101", # Using assert
  "PLR2004", # magic-value-comparison
  "D107", # Missing docstring in `__init__`
  "PLW2901", # redefined-loop-name
  "D202", # No blank lines allowed after function docstring
  "F811" # redefinition of function
]
